refactor: change style component to root auth folder
