--- web-dashboard/src/pages/auth/signup/index.module.less
@@ -1,23 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    margin: 6rem auto;
-    max-width: 500px;
-    height: 100%;
-}
-
-.content {
-    padding: 1rem;
-}
-
-.content_title {
-    text-align: center;
-}
-
-.content_body {
-    padding-top: 2.5rem;
-}
-
-.text_center {
-    text-align: center;
-}
